refac: put the styles from the stages into the theme, put the stepper into the components
