+++ src/theme/palette.ts
@@ -46,10 +46,10 @@ const INFO = {
 const SUCCESS = {
   lighter: '#E9FCD4',
   light: '#AAF27F',
-  main: '#54D62C',
+  main: '#18A670',
   dark: '#229A16',
   darker: '#08660D',
-  contrastText: GREY[800],
+  contrastText: '#fff',
 };
 
 const WARNING = {
